Remove GitHub contributors from authors in plugin.yml
This was intended to give credit to contributors but I didn't consider the fact that all names will always be displayed in the console whenever the server wants the admin to contact the plugin author, which may have been misleading at times. Contributors can still be found on GitHub's own list.